src/data-ingestor/requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
kafka-python==2.0.2
requests==2.32.3
